+++ CHANGELOG.rst
@@ -2,6 +2,11 @@ Changelog
 ---------
 
 v2.2 (??? 2024)
+    * Access positional element of Dataframe with iloc in tests.
+    * Changed deprecated numpy.in1d to numpy.isin in `fsread`.
+    * Delete test for float128 in `ncio` because no 128-bit datatype
+      in netcdf.
+    * Assure fill_value is string in `sread`.
     * Argument 'parents' to `mcPlot` to pass extra command line
       arguments.
     * Added upper keyword in `abc2plot`.